Broke ControlledTierCard out of DAR.
